requirements.txt
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfminer.six>=20221105
python-pptx>=0.6.21
openpyxl>=3.1.0
torch>=2.0.0
rapidfuzz>=3.0.0
# Phase 2 dependencies
chromadb>=0.4.22
sentence-transformers>=2.3.0
transformers>=4.36.0
tqdm>=4.66.0
numpy>=1.24.0